refactor(heroRoller): move /heroroll into new subsystem
